allow go install github.com/flashbots/mev-boost (#214)
* move cli entrypoint to cli/

* add new root entrypoint

* added /config to fix version baking
