+++ cli/main.go
@@ -1,0 +1,158 @@
+package cli
+
+import (
+	"flag"
+	"fmt"
+	"os"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/flashbots/mev-boost/config"
+	"github.com/flashbots/mev-boost/server"
+	"github.com/sirupsen/logrus"
+)
+
+const (
+	genesisForkVersionMainnet = "0x00000000"
+	genesisForkVersionKiln    = "0x70000069"
+	genesisForkVersionRopsten = "0x80000069"
+	genesisForkVersionSepolia = "0x90000069"
+)
+
+var (
+	// defaults
+	defaultLogJSON            = os.Getenv("LOG_JSON") != ""
+	defaultLogLevel           = getEnv("LOG_LEVEL", "info")
+	defaultListenAddr         = getEnv("BOOST_LISTEN_ADDR", "localhost:18550")
+	defaultRelayTimeoutMs     = getEnvInt("RELAY_TIMEOUT_MS", 2000) // timeout for all the requests to the relay
+	defaultRelayCheck         = os.Getenv("RELAY_STARTUP_CHECK") != ""
+	defaultGenesisForkVersion = getEnv("GENESIS_FORK_VERSION", "")
+	maxHeaderBytes            = getEnvInt("MAX_HEADER_BYTES", 4000) // max header byte size for requests for dos prevention
+
+	// cli flags
+	printVersion = flag.Bool("version", false, "only print version")
+	logJSON      = flag.Bool("json", defaultLogJSON, "log in JSON format instead of text")
+	logLevel     = flag.String("loglevel", defaultLogLevel, "log-level: trace, debug, info, warn/warning, error, fatal, panic")
+
+	listenAddr     = flag.String("addr", defaultListenAddr, "listen-address for mev-boost server")
+	relayURLs      = flag.String("relays", "", "relay urls - single entry or comma-separated list (scheme://pubkey@host)")
+	relayTimeoutMs = flag.Int("request-timeout", defaultRelayTimeoutMs, "timeout for requests to a relay [ms]")
+	relayCheck     = flag.Bool("relay-check", defaultRelayCheck, "check relay status on startup and on the status API call")
+
+	// helpers
+	useGenesisForkVersionMainnet = flag.Bool("mainnet", false, "use Mainnet genesis fork version 0x00000000 (for signature validation)")
+	useGenesisForkVersionKiln    = flag.Bool("kiln", false, "use Kiln genesis fork version 0x70000069 (for signature validation)")
+	useGenesisForkVersionRopsten = flag.Bool("ropsten", false, "use Ropsten genesis fork version 0x80000069 (for signature validation)")
+	useGenesisForkVersionSepolia = flag.Bool("sepolia", false, "use Sepolia genesis fork version 0x90000069 (for signature validation)")
+	useCustomGenesisForkVersion  = flag.String("genesis-fork-version", defaultGenesisForkVersion, "use a custom genesis fork version (for signature validation)")
+)
+
+var log = logrus.WithField("module", "cmd/mev-boost")
+
+// Main starts the mev-boost cli
+func Main() {
+	flag.Parse()
+	logrus.SetOutput(os.Stdout)
+
+	if *printVersion {
+		fmt.Printf("mev-boost %s\n", config.Version)
+		return
+	}
+
+	if *logJSON {
+		log.Logger.SetFormatter(&logrus.JSONFormatter{})
+	} else {
+		log.Logger.SetFormatter(&logrus.TextFormatter{
+			FullTimestamp: true,
+		})
+
+	}
+
+	if *logLevel != "" {
+		lvl, err := logrus.ParseLevel(*logLevel)
+		if err != nil {
+			log.Fatalf("Invalid loglevel: %s", *logLevel)
+		}
+		logrus.SetLevel(lvl)
+	}
+
+	log.Infof("mev-boost %s", config.Version)
+
+	genesisForkVersionHex := ""
+	if *useCustomGenesisForkVersion != "" {
+		genesisForkVersionHex = *useCustomGenesisForkVersion
+	} else if *useGenesisForkVersionMainnet {
+		genesisForkVersionHex = genesisForkVersionMainnet
+	} else if *useGenesisForkVersionKiln {
+		genesisForkVersionHex = genesisForkVersionKiln
+	} else if *useGenesisForkVersionRopsten {
+		genesisForkVersionHex = genesisForkVersionRopsten
+	} else if *useGenesisForkVersionSepolia {
+		genesisForkVersionHex = genesisForkVersionSepolia
+	} else {
+		log.Fatal("Please specify a genesis fork version (eg. -mainnet / -kiln / -ropsten / -sepolia / -genesis-fork-version flags)")
+	}
+	log.Infof("Using genesis fork version: %s", genesisForkVersionHex)
+
+	relays := parseRelayURLs(*relayURLs)
+	if len(relays) == 0 {
+		log.Fatal("No relays specified")
+	}
+	log.WithField("relays", relays).Infof("using %d relays", len(relays))
+
+	relayTimeout := time.Duration(*relayTimeoutMs) * time.Millisecond
+	if relayTimeout <= 0 {
+		log.Fatal("Please specify a relay timeout greater than 0")
+	}
+
+	opts := server.BoostServiceOpts{
+		Log:                   log,
+		ListenAddr:            *listenAddr,
+		Relays:                relays,
+		GenesisForkVersionHex: genesisForkVersionHex,
+		RelayRequestTimeout:   relayTimeout,
+		RelayCheck:            *relayCheck,
+		MaxHeaderBytes:        maxHeaderBytes,
+	}
+	server, err := server.NewBoostService(opts)
+	if err != nil {
+		log.WithError(err).Fatal("failed creating the server")
+	}
+
+	if *relayCheck && !server.CheckRelays() {
+		log.Fatal("no relay available")
+	}
+
+	log.Println("listening on", *listenAddr)
+	log.Fatal(server.StartHTTPServer())
+}
+
+func getEnv(key string, defaultValue string) string {
+	if value, ok := os.LookupEnv(key); ok {
+		return value
+	}
+	return defaultValue
+}
+
+func getEnvInt(key string, defaultValue int) int {
+	if value, ok := os.LookupEnv(key); ok {
+		val, err := strconv.Atoi(value)
+		if err == nil {
+			return val
+		}
+	}
+	return defaultValue
+}
+
+func parseRelayURLs(relayURLs string) []server.RelayEntry {
+	ret := []server.RelayEntry{}
+	for _, entry := range strings.Split(relayURLs, ",") {
+		relay, err := server.NewRelayEntry(entry)
+		if err != nil {
+			log.WithError(err).WithField("relayURL", entry).Fatal("Invalid relay URL")
+		}
+		ret = append(ret, relay)
+	}
+	return ret
+}
