+++ config/vars.go
@@ -1,0 +1,7 @@
+package config
+
+// Version is the version of the software, set at build time
+var Version = "dev"
+
+// BuildTime is the output of `date` at build time. Eg. "Sat Jul 16 13:18:55 CEST 2022"
+var BuildTime string
