+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/flashbots/mev-boost/cli"
+)
+
+func main() {
+	cli.Main()
+}
